EES-925 Move SubscriptionForm into SubscriptionPage
